refactor: move some methods to separate file, as functions; clean up some things
